backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brlapi==0.8.5
Brotli==1.1.0
cachetools==6.2.6
catfish==4.16.4
ccsm==0.9.14.2
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
colorama==0.4.6
command-not-found==0.3
compizconfig-python==0.9.14.2
configobj==5.0.8
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
eyed3==0.9.7
fastapi==0.135.3
filelock==3.13.1
filetype==1.2.0
fsspec==2026.3.0
google-auth==2.49.1
google-genai==1.71.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html2text==2024.2.26
httpcore==1.0.9
httplib2==0.20.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.6
ifaddr==0.2.0
IMDbPY==2021.4.18
jiter==0.13.0
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libevdev==0.5
louis==3.29.0
lxml==5.2.1
Mako==1.3.2.dev0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
menulibre==2.4.0
mmh3==5.2.1
mugshot==0.4.3
multidict==6.7.1
multiprocess==0.70.16
mutagen==1.46.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
onboard==1.4.1
openai==2.31.0
packaging==24.0
pexpect==4.9.0
pillow==10.2.0
postgrest==2.28.3
propcache==0.4.1
psutil==5.9.8
ptyprocess==0.7.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycairo==1.25.1
pycryptodomex==3.20.0
pycups==2.0.1
pycurl==7.45.3
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pyelftools==0.30
Pygments==2.17.2
PyGObject==3.48.2
pyiceberg==0.11.1
PyICU==2.12
PyJWT==2.12.1
PyMuPDF==1.27.2.2
PyNaCl==1.5.0
pyparsing==3.1.1
pyparted==3.12.0
pypng==0.20231004.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
pyroaring==1.0.4
python-apt==2.7.7+ubuntu5.2
python-dateutil==2.9.0.post0
python-debian==0.1.49+ubuntu2
python-dotenv==1.2.2
python-gnupg==0.5.2
python-multipart==0.0.24
pyudev==0.24.0
pyxdg==0.28
PyYAML==6.0.1
qrcode==7.4.2
realtime==2.28.3
repolib==2.2.1
requests==2.31.0
requests-file==1.5.1
rich==13.7.1
setproctitle==1.3.3
setuptools==68.1.2
six==1.16.0
sniffio==1.3.1
fastapi
uvicorn[standard]
SpeechRecognition
pydantic
google-generativeai
python-dotenv==3.16.0
soupsieve==2.5
SpeechRecognition==3.16.0
starlette==1.0.0
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
systemd-python==235
tenacity==9.1.4
tldextract==3.1.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ufw==0.36.2
Unidecode==1.3.8
urllib3==2.0.7
urwid==2.6.10
uvicorn==0.44.0
wadllib==1.3.6
wcwidth==0.2.5
websockets==15.0.1
xdg==5
xkit==0.0.0
yarl==1.23.0
yt-dlp==2024.4.9
zstandard==0.25.0
